Removed fix status
The "fix status" was supposed to provide to MiMa users information on whether a
binary re-writing tool (which does not yet exist) would have been able to
automagically patch the binary incompatibility.

Unfortunately, the above mentioned re-writing tool does not yet exist, and
therefore the reported information is currently useless (if not confusing) for
the user.