Fixed failing test
The issue was that the test runner (i.e., ``CollectProblemsTest``), had an
hardcoded reference to the Scala library, which is needed to correctly run the
test. However, due to recent git-history re-writing (which has considerably
reduced the repo's size), the Scala library JAR referenced in the test was no
longer available.

The fix is to pass the ``testClasspath`` to the test runner, and at the moment
the ``testClasspath`` only contains a reference to the Scala library's path
used to build the project itself.