Added support for parsing classfile version 51 (JDK7). Fix #66
I'm surprised that this issue was discovered only now, because we were missing
support for parsing JDK7 classfile (!) - so not just JDK8 ones.

The fix was to simply extend the `ClassfileParser` to consider the additional
constants that may appear in the classfile's Constant Pool.  Because the
implementation of `ClassfileParser` was copied from scalac, we simply applied
here the same fix as in
https://github.com/scala/scala/commit/e78896f0ab7240e9bdcd98c51cbf6bc0ea277950#diff-400a2e6482bc824c3840c10ecc43e7b0R140.
We should definitely work on removing this duplication (I'd say it's here for
historical reasons), but doing so is out of the scope of this PR.