Adding an abstract method is a binary compatible evolution. Fix #69
As stated in Java binary compatibility doc
(https://docs.oracle.com/javase/specs/jls/se8/html/jls-13.html),
"adding new fields, methods, or constructors to an existing class or
interface" is a binary compatible evolution.

Adding an abstract method to an interface or class is probably the
most surprising binary compatible evolution, but it makes a lot of sense
and it's pretty neat.

Also, by fixing this problem, we now also fully support Java8 default methods.
Which is why this commit also Fix #68.