only run the test suite on Java 8
reporter/functional-tests/src/test/java8-turning-interface-method-into-default-method
requires Java 8.  we could do some build trickery to exclude that test
depending on the Java version... but it doesn't seem worth the effort to me.
I think it is sufficient these days to just test on Java 8, on a project like
MiMa that is only used in builds, not at runtime.