in build, set name of root project
allowing it to default to "root" leads to this error in the
community build, after extraction:
[error] Fatal: multiple projects have the same artifacts visible in the same space.
[error]   root#root  from scala-records and mima, both visible in space "default"

perhaps dbuild ought to notice that `publish` and `publishLocal` are
`()` in the root project, but it doesn't (not dbuild 0.9.1 anyway, I
don't know about later versions).  and as far as I can see, setting
the name is harmless.