updated expected test results for 2.12.0-RC2 changes
fixes com.typesafe.tools.mima.lib.TestFailed on
test-class-method-abstract-override-of-concrete-supertrait-method-nok

conferred with Adriaan on this. it looks like this is because of the
restoration of mixin forwarders in RC2. so we just update the expected
output to match the new behavior. (this makes the 2.12 expected
output the same as the 2.11 output again, so we just remove the
2.12-specific file.)