Do not skip classes with symbolic names in objects
We used to skip all classes whose bytecode name contains “$$” under the
assumption that those would be names with synthetic components like
“$$anonfun” but these names can also occur when a class name starting
with a symbolic character is nested inside an object (e.g. “A$$less” is
a class named “<“ inside an “object A”). Instead we need to look at the
decoded names. They still contain synthetic components such as
“$$anonfun” verbatim but no accidental “$$” sequences.

Fixes #158.