Allow multiple exclude files per version to be loaded from a directory
So, instead of creating a single

mima-filters/10.1.5.backwards.excludes

file to which everyone appends, you can now create a directory with the same name
containing multiple exclude files per change like that:

mima-filters/10.1.5.backwards.excludes/pr1234.excludes
mima-filters/10.1.5.backwards.excludes/pr5678.excludes

etc.

This will prevent the common merge conflicts when everyone tries to append
to the same file when in fact the ordering of entries is not interesting.