Verify all overloads for result types & signatures
Instead of taking the first method which "matches type" (which actually
means "has the same parameters") to check whether the result types also
match, check them all.

The impact is visible in the integration test changes: ZipArchive's
underlyingSource is "just" a (generic) signature problem, not a result
type problem.  Specifically it still returns a scala.Some, except it's a:

> ()Lscala/Some<Lscala/reflect/io/AbstractFile;>; rather than ()Lscala/Some<Lscala/reflect/io/ZipArchive;>;