downgrade ScalaTest to 3.0.x, for community build
I think we had moved to 3.1.x because there was an interval when
3.0.x wasn't available for whatever Scala 2.13 prerelease is current
at the time. that consideration no longer applies.

in the community build, we're still on ScalaTest 3.0 (since 3.1 isn't
final yet), so I'd rather do the downgrade here for now, rather than
have to freeze mima in the community build (which is what we've done
with some other projects that have moved to 3.1)