Drop building in Scala 2.13
This drops the setup & testing of MiMa using Scala 2.13.

The reason is that there is no need for this.  MiMa is by-and-large
consumed via its sbt plugins, which is not 1.x only.  sbt 1.x will
always be Scala 2.12 only, and there are not real plans for sbt 2.x.
Therefore we can stop building with Scala 2.13.

Note that MiMa continues to support verifying the binary compatibility
of Scala 2.11-2.13 projects - that support is crucial and not going
anywhere.