Reimplement problem discovery for dropped overload
In #362 I changed the logic so that when an overloaded is dropped it
doesn't get reported as an "incompatible method type" or "incompatible
result type" problem.

Looking back the logic is quite short-sighted.

So here I attempt to make it more straight-forward: if there were
overloads, then the method is missing - it's not incompatible.