Report the loss of trait init methods
In the encoding of traits the presence of a method with a body (a
"concrete method") triggers the generation of a trait initialisation
method, "$init$", which will be invoked by subclasses during
construction.  The effect this has is that dropping the last concrete
method(s) from a trait _removes_ the trait init method, which is a
binary incompatible change to client's subclasses.

Fixes #426