Move publishTo to build-setting
This is to fix sbt-plugin publishing to Maven Central.

Maybe the apiSettings should move there too.  Maybe mimaSettings should
move there too.  But then I'd have to check that that scope is actually
read, & I'd have to check that projects that the opt out of mima still
opt out.  For another day.