Drop a redundant Travis CI job
Previously the setup created two "sbt ++$TRAVIS_SCALA test" build jobs,
for openjdk8 and openjdk11, and then the rest of the jobs in the "jobs"
section (once, on openjdk8 - because it's the first jdk value).

Now it will run the whole test/mimaReportBinaryIssues/etc job once on
Java 8 and once on Java 11, and then testFunctional/etc on Java 8 (like
before).